Add failing get() and delete() tests
I want to figure out a way to do this that doesn't require calling
self.test_save(). Saving obviously can't go in setUp, because then if
saving would fail, then setUp wouldn't work, and setUp isn't a test, so...
Maybe manual insertion of a record in the database?

Also, some other tests are getting errors during the test. The cause of
this is Roll not allowing negative modifiers. Gonna have to fix that and
add tests for it!