Add Text class
This class loads and provides access to the text files used to store user
messages. I just added it because it was easy to do and makes main.py
marginally prettier.

I guess this is also a good time to justify the decision to keep this text
in separate files. I just don't like the look of large strings written
directly in code. This is the reason for both this functionality, and the
functionality of loading SQL queries from text files, though they differ a
bit in what exactly they do.